+++ tapis-systemsapi/src/main/java/edu/utexas/tacc/tapis/systems/api/requests/ReqPostSystem.java
@@ -63,4 +63,5 @@ public final class ReqPostSystem
   public List<Capability> jobCapabilities;
   public String[] tags = EMPTY_STR_ARRAY;
   public Object notes = DEFAULT_NOTES;
+  public String importRefId;
 }