+++ tapis-systemsapi/src/main/java/edu/utexas/tacc/tapis/systems/api/requests/ReqPutSystem.java
@@ -50,4 +50,5 @@ public final class ReqPutSystem
   public List<Capability> jobCapabilities;
   public String[] tags = EMPTY_STR_ARRAY;
   public Object notes = DEFAULT_NOTES;
+  public String importRefId;
 }