+++ tapis-systemslib/src/main/resources/db/migration/V001__init.sql
@@ -76,6 +76,7 @@ CREATE TABLE systems
   job_capabilities JSONB NOT NULL,
   tags       TEXT[] NOT NULL,
   notes      JSONB NOT NULL,
+  import_ref_id TEXT,
   uuid uuid NOT NULL,
   deleted    BOOLEAN NOT NULL DEFAULT false,
   created    TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT (NOW() AT TIME ZONE 'utc'),
@@ -120,6 +121,7 @@ COMMENT ON COLUMN systems.batch_scheduler_profile IS 'Scheduler profile for the 
 COMMENT ON COLUMN systems.job_capabilities IS 'Capabilities associated with system';
 COMMENT ON COLUMN systems.tags IS 'Tags for user supplied key:value pairs';
 COMMENT ON COLUMN systems.notes IS 'Notes for general information stored as JSON';
+COMMENT ON COLUMN systems.import_ref_id IS 'Reference for systems created via import';
 COMMENT ON COLUMN systems.deleted IS 'Indicates if system has been soft deleted';
 COMMENT ON COLUMN systems.created IS 'UTC time for when record was created';
 COMMENT ON COLUMN systems.updated IS 'UTC time for when record was last updated';